Parameter Make_legacy.1-I

type index

type context

type tree

type hash

type contents := bytes

type step := string

type commit_info

type batch

val batch : index -> (batch -> 'a Lwt.t) -> 'a Lwt.t

val commit_info_encoding : commit_info Tezos_base__TzPervasives.Data_encoding.t

val hash_encoding : hash Tezos_base__TzPervasives.Data_encoding.t

val hash_equal : hash -> hash -> bool

module Block_header : sig ... end

module Pruned_block : sig ... end

module Block_data : sig ... end

module Protocol_data : sig ... end

module Commit_hash : sig ... end

val context_parents : context -> Commit_hash.t list

val context_info : context -> commit_info

val get_context : index -> Block_header.t -> context option Lwt.t

val set_context : info:commit_info -> parents:Commit_hash.t list -> context -> Block_header.t -> bool Lwt.t

val context_tree : context -> tree

val tree_iteri_unique : (int -> [ `Branch of (step * hash) list | `Leaf of contents ] -> unit Lwt.t) -> tree -> unit Lwt.t

Visit each branch and leaf of the given tree exactly once, in depth-first post-order traversal. Branch children are visited in ascending key order. Memory usage is linear in the size of the tree.

val make_context : index -> context

val update_context : context -> tree -> context

val add_bytes : batch -> bytes -> tree Lwt.t

val add_dir : batch -> (step * hash) list -> tree option Lwt.t